font,vg: migrate vg.Length to font.Length
