+++ font/doc.go
@@ -1,0 +1,6 @@
+// Copyright ©2021 The Gonum Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+// Package font provides types to describe and select text font faces.
+package font // import "gonum.org/v1/plot/font"
